oops, needed to cast to uchar
git-svn-id: file:///home/justin/tmp/tmp/svn/cutestuff/trunk@37 65d0b1b9-db85-4114-8a31-8ae23c1c3140

+++ util/sha1.cpp
@@ -184,7 +184,7 @@ QString SHA1::digest(const QString &in)
 	QString out;
 	for(int n = 0; n < (int)a.size(); ++n) {
 		QString str;
-		str.sprintf("%02x", a[n]);
+		str.sprintf("%02x", (uchar)a[n]);
 		out.append(str);
 	}
